third_party/WebKit/LayoutTests/http/tests/images/image-currentsrc-load-expected.txt

PASS img.currentSrc is ""
PASS img.currentSrc is expectedCurrentSrc
PASS successfullyParsed is true

TEST COMPLETE
